RCPS-106 Student Information Management: Rename backend app package
